Extracted image url loader into a separate class.
