+++ app/src/main/java/com/android/academy/fundamentals/homework/presentation/features/main/MainActivity.kt
@@ -3,21 +3,22 @@ package com.android.academy.fundamentals.homework.presentation.features.main
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.android.academy.fundamentals.homework.R
+import com.android.academy.fundamentals.homework.data.MovieRepositoryImpl
+import com.android.academy.fundamentals.homework.data.remote.retrofit.ImageUrlLoader
 import com.android.academy.fundamentals.homework.data.remote.retrofit.RetrofitDataSource
 import com.android.academy.fundamentals.homework.di.MovieRepositoryProvider
 import com.android.academy.fundamentals.homework.di.NetworkModule
+import com.android.academy.fundamentals.homework.domain.MovieRepository
 import com.android.academy.fundamentals.homework.presentation.features.moviedetails.view.MovieDetailsFragment
 import com.android.academy.fundamentals.homework.presentation.features.movies.view.MoviesListFragment
-import com.android.academy.fundamentals.homework.domain.MovieRepository
-import com.android.academy.fundamentals.homework.data.MovieRepositoryImpl
 
 internal class MainActivity : AppCompatActivity(),
-                     MoviesListFragment.MoviesListItemClickListener,
-                     MovieDetailsFragment.MovieDetailsBackClickListener,
-                     MovieRepositoryProvider {
+    MoviesListFragment.MoviesListItemClickListener,
+    MovieDetailsFragment.MovieDetailsBackClickListener,
+    MovieRepositoryProvider {
 
     private val networkModule = NetworkModule()
-    private val remoteDataSource = RetrofitDataSource(networkModule.api)
+    private val remoteDataSource = RetrofitDataSource(networkModule.api, ImageUrlLoader(networkModule.api))
     private val movieRepository = MovieRepositoryImpl(remoteDataSource)
 
     override fun onCreate(savedInstanceState: Bundle?) {
